Detach Keystone to run app
Babel is trying to load .less files. Keystone wasn't passing an instance of express to the routes. Going forward it may be best to implement all database queries without using Keystone's helpers. In addition to separating the controllers from Keystone, and express instance needs to be made at the very least more accessable with the CMS/admin routing becoming secondary.